backend/requirements.txt
fastapi
uvicorn
pymupdf
docx2txt
scikit-learn
